+++ Chasm.Formatting/CHANGELOG.md
@@ -6,6 +6,7 @@
 - ✨ Added `FormattingExtensions.TryCopyTo<T>(this Span<T>, Span<T>, out int)`;
 - ✨ Added `FormattingExtensions.TryCopyTo<T>(this ReadOnlySpan<T>, Span<T>, out int)`;
 - 🧑‍💻 Improved debugger displays of `SpanParser` and `SpanBuilder`;
+- ⚡️ Removed unnecessary cast to long in `SpanBuilder.Append(int)`;
 
 ### v2.2.1
 - 🚑️ Removed `SpanBuilder.TryFormat(ISpanBuildableFormat, ReadOnlySpan<char>, Span<char>, out int)`;